ci(bench): refine bcs serialized size bench by passing function as input values (#1568)
This commit refines the benchmarks for 'bcs serialized size' by making the benchmark function take function values as inputs. This is done by redefining the benchmark function to take a vector of function containers as parameter, which allows different evaluation functions to be passed in as arguments. It provides a more flexible way to perform benchmarks and measure different aspects of the encoding and serializing processes.