ci(bench): refactor benchmark configuration and update related files (#1574)
1. The tx_exec_benchmark function has been moved from the `tx` module to a new `tx_exec` module. 

This included updating import paths in both the function's definition and where it is used, in `bench_tx_exec.rs` and `bench_tx_exec_blk.rs`. The commit improves the organization of the codebase by placing each major functionality in its own module.

2. Remove transaction query benchmarks

The benchmark files `bench_tx.rs` and `bench_tx_query.rs` have been deleted as part of code cleanup. This removal is reflected in the project configuration file, `Cargo.toml` as well, thereby ensuring that these benchmarks are no longer part of the build process.

3.  refactor benchmark configuration and update related files

This commit refactors the configuration of the benchmarking function, moving the setup code into a new `config.rs` file. The main focus of this commit is to streamline the configuration of benchmarks, making it easier to adjust settings for different benchmarks. It introduces a new `BenchTxConfig` struct for defining the benchmark configuration and uses it to replace the previous use of environment variables. This commit also modifies various files to match the new configuration method.