feat(common): add new 'rooch-common' crate with utility functions (#2067)
This commit introduces a new 'rooch-common' crate to the workspace and that contains utility functions for the project. It includes a function for converting byte sizes to human-readable formats, which has been incorporated into the ORD update code. This will allow for clearer log messages describing the size of the batches processed.