fix(build): try to fix build issue on MSVC and docker caused by using jemalloc (#2135)
* fix(build): fix build failed on msvc

* chore: clean unused dependencies

* add jemalloc back for rooch cmd

* rm local config for cargo config

* refine deps

rust-rocksdb -> rocksdb: too few people using, change back. and disable jemalloc feature

tikv-jemallocator -> jemallocator: same crate with tidy name

* Update Dockerfile dependencies and Rust version

Added 'make' to the Dockerfile dependencies and updated the Rust version from 1.70.0 to 1.78.0. These changes are crucial for ensuring up-to-date builds and compatibility with other modern tools.

* Remove 'polyval' and 'portable-atomic' packages

The 'polyval' and 'portable-atomic' packages have been removed from the Cargo.lock file. This deletion reduces the number of dependencies and simplifies the project structure.

* docs: add 'make' to installation dependencies

The 'make' package was missing from the installation dependencies in the Getting Started guides for both the English and Chinese versions. This has been corrected in the documentation.