Add performance tracking to update batches
A start time variable has been added at the beginning of processing loops in both genesis_utxo.rs and genesis_ord.rs files. Additionally, an informative print statement has been added after the processing to display the number of updates in the batch and the elapsed processing time. This will aid in tracking and improving the performance of update batches.