Refactor statedb export code in rooch
The commit refactors the statedb export code in the rooch crate. The notable changes include the removal of redundant and commented code, as well as the alteration of the logging and writing methods to improve efficiency. In detail, unnecessary print statements have been subtracted and log::info is now utilized instead. The CSV records are now written in a more direct manner.