Implement FileCacheManager in import command
A FileCacheManager has been added to the import command in order to optimize reading and handling large files. The code now takes into account the size of the read data, dropping cache when it exceeds the limit. Error handling was also converted from using Result returns to unwrapping for simplicity.